requirements.txt
asgiref==3.3.1
autopep8==1.5.4
dj-database-url==0.5.0
Django==3.1.4
django-heroku==0.3.1
djangorestframework==3.12.2
gunicorn==20.0.4
psycopg2==2.8.6
pycodestyle==2.6.0
pytz==2020.5
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
